Произведена реорганизация файловой структуры и частичный рефакторинг
