+++ source/pkcs_15_cryptographic_token/kc_includes.h
@@ -1,6 +1,8 @@
 #ifndef ASN_1_CONVERSION_INCLUDES_H
 #define ASN_1_CONVERSION_INCLUDES_H
 
+//todo: move to libarkypt.h
+
 #include <stdio.h>
 #include <ctype.h>
 #include <stdint.h>
